Make SimpleLogger a custom element, move some code
Custom elements on the webpage are now defined in the components folder.
