--- SimpleLogger.css
@@ -1,47 +1,0 @@
-.simpleLogger {
-	margin: 15px;
-	color: white;
-	font-family: monospace;
-	font-size: 14px;
-	max-height: 300px;
-	overflow: auto;
-	text-align: left;
-	position: relative;
-}
-.logHeader {
-	padding: 2px 12px;
-	position: sticky;
-	top: 0;
-	background: #556;
-}
-.logHeader * {
-	margin-left: 30px;
-}
-.logHeader button {
-	background: inherit;
-	border: 1px solid #CCC;
-	color: white;
-}
-.log {
-	margin: 0;
-	padding: 2px 12px;
-	background: #223;
-}
-.log:nth-child(2n) {
-	background: #334;
-}
-.warning > .logText::before {
-	content: "\26A0\FE0F ";
-}
-.error > .logText::before {
-	content: "\1F6A8 ";
-}
-.info > .logText::before {
-	content: "\2139\FE0F ";
-}
-.timestamp {
-	margin-right: 5px;
-	padding: 0 2px;
-	background: #556;
-	color: #CCC;
-}
