+++ index.html
@@ -175,7 +175,7 @@
 					</fieldset>
 				</form>
 				<div id="completedPacksCont"></div>
-				<div id="log"></div>
+				<simple-logger id="log"></simple-logger>
 			</div>
 			<div id="languageSelectorCont" title="Change language" data-translate-title="change_language">
 				<div id="languageSelectorButton"></div>
